- refactor into new NetworkRestApi service
