+++ http-api/src/main/java/bisq/http_api/rest_api/RestApiResourceConfig.java
@@ -2,6 +2,7 @@ package bisq.http_api.rest_api;
 
 import bisq.http_api.rest_api.domain.explorer.ExplorerRestApi;
 import bisq.http_api.rest_api.domain.market_price.MarketPriceRestApi;
+import bisq.http_api.rest_api.domain.network.NetworkRestApi;
 import bisq.http_api.rest_api.domain.offers.OfferbookRestApi;
 import bisq.http_api.rest_api.domain.settings.SettingsRestApi;
 import bisq.http_api.rest_api.domain.trades.TradeRestApi;
@@ -16,6 +17,7 @@ public class RestApiResourceConfig extends BaseRestApiResourceConfig {
     public RestApiResourceConfig(String swaggerBaseUrl,
                                  OfferbookRestApi offerbookRestApi,
                                  TradeRestApi tradeRestApi,
+                                 NetworkRestApi networkRestApi ,
                                  UserIdentityRestApi userIdentityRestApi ,
                                  MarketPriceRestApi marketPriceRestApi,
                                  SettingsRestApi settingsRestApi,
@@ -29,6 +31,7 @@ public class RestApiResourceConfig extends BaseRestApiResourceConfig {
         // with AbstractBinder to register resources as classes for Swagger
         register(OfferbookRestApi.class);
         register(TradeRestApi.class);
+        register(NetworkRestApi.class);
         register(UserIdentityRestApi.class);
         register(MarketPriceRestApi.class);
         register(SettingsRestApi.class);
@@ -39,6 +42,7 @@ public class RestApiResourceConfig extends BaseRestApiResourceConfig {
             protected void configure() {
                 bind(offerbookRestApi).to(OfferbookRestApi.class);
                 bind(tradeRestApi).to(TradeRestApi.class);
+                bind(networkRestApi).to(NetworkRestApi.class);
                 bind(userIdentityRestApi).to(UserIdentityRestApi.class);
                 bind(marketPriceRestApi).to(MarketPriceRestApi.class);
                 bind(settingsRestApi).to(SettingsRestApi.class);
