+++ http-api/src/main/java/bisq/http_api/web_socket/WebSocketRestApiResourceConfig.java
@@ -3,6 +3,7 @@ package bisq.http_api.web_socket;
 import bisq.http_api.rest_api.RestApiResourceConfig;
 import bisq.http_api.rest_api.domain.explorer.ExplorerRestApi;
 import bisq.http_api.rest_api.domain.market_price.MarketPriceRestApi;
+import bisq.http_api.rest_api.domain.network.NetworkRestApi;
 import bisq.http_api.rest_api.domain.offers.OfferbookRestApi;
 import bisq.http_api.rest_api.domain.settings.SettingsRestApi;
 import bisq.http_api.rest_api.domain.trades.TradeRestApi;
@@ -16,10 +17,11 @@ public class WebSocketRestApiResourceConfig extends RestApiResourceConfig {
     public WebSocketRestApiResourceConfig(String swaggerBaseUrl,
                                           OfferbookRestApi offerbookRestApi,
                                           TradeRestApi tradeRestApi,
+                                          NetworkRestApi networkRestApi,
                                           UserIdentityRestApi userIdentityRestApi,
                                           MarketPriceRestApi marketPriceRestApi,
                                           SettingsRestApi settingsRestApi,
                                           ExplorerRestApi explorerRestApi) {
-        super(swaggerBaseUrl, offerbookRestApi, tradeRestApi, userIdentityRestApi, marketPriceRestApi, settingsRestApi, explorerRestApi);
+        super(swaggerBaseUrl, offerbookRestApi, tradeRestApi, networkRestApi, userIdentityRestApi, marketPriceRestApi, settingsRestApi, explorerRestApi);
     }
 }
